Update the document definition to be less restrictive.
I've reduced the severity of a document standard violation to "undefined 
behavior" and a warning "may" be emitted.

The primary reason is that we want to avoid the need to check each and 
every document, which may be computationally very expensive and the 
strict need to reject all documents that do not adhere to the standard.

Instead it is clear that non-standard documents are discouraged, this 
might lead to warnings and generally undefined behavior.