Made mention that we are comparing to ideal gas law
We compare our simulation volume to the volume calculated by the ideal
gas law, even though we are not simulating an ideal gas. This was
confusing since users noticed we were not simulating an ideal gas. We
added more language to clarify that we are NOT simulating an ideal gas
but we are comparing the simulation volume to an ideal gas.